# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Swim Coach API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /coach/bookings:
    get:
      operationId: getCoachBookings
      tags: [coach]
      summary: Get all bookings for coach review
      parameters:
        - in: header
          name: x-coach-password
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of bookings sorted newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
        "401":
          description: Unauthorized
  /coach/bookings/{id}/approve:
    patch:
      operationId: approveBooking
      tags: [coach]
      summary: Approve a booking and auto-create client
      parameters:
        - in: header
          name: x-coach-password
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Booking not found
  /coach/bookings/{id}/reject:
    patch:
      operationId: rejectBooking
      tags: [coach]
      summary: Reject a booking with optional note
      parameters:
        - in: header
          name: x-coach-password
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                note:
                  type: string
      responses:
        "200":
          description: Booking rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Booking not found
  /coach/clients:
    get:
      operationId: getCoachClients
      tags: [coach]
      summary: List all clients
      parameters:
        - in: header
          name: x-coach-password
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
  /coach/clients/{id}:
    patch:
      operationId: updateCoachClient
      tags: [coach]
      summary: Update client notes or status
      parameters:
        - in: header
          name: x-coach-password
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                notes:
                  type: string
                status:
                  type: string
      responses:
        "200":
          description: Client updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /coach/plans:
    get:
      operationId: getCoachPlans
      tags: [coach]
      summary: List all coaching plans
      parameters:
        - in: header
          name: x-coach-password
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of plans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CoachingPlan"
    post:
      operationId: createCoachPlan
      tags: [coach]
      summary: Create a new coaching plan
      parameters:
        - in: header
          name: x-coach-password
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlanRequest"
      responses:
        "201":
          description: Plan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachingPlan"
  /coach/plans/{id}:
    patch:
      operationId: updateCoachPlan
      tags: [coach]
      summary: Update a coaching plan
      parameters:
        - in: header
          name: x-coach-password
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePlanRequest"
      responses:
        "200":
          description: Plan updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoachingPlan"
  /coach/sessions:
    get:
      operationId: getCoachSessions
      tags: [coach]
      summary: List all sessions
      parameters:
        - in: header
          name: x-coach-password
          required: true
          schema:
            type: string
        - in: query
          name: clientId
          schema:
            type: integer
      responses:
        "200":
          description: List of sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Session"
    post:
      operationId: createCoachSession
      tags: [coach]
      summary: Create a new session
      parameters:
        - in: header
          name: x-coach-password
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSessionRequest"
      responses:
        "201":
          description: Session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
  /coach/sessions/{id}:
    patch:
      operationId: updateCoachSession
      tags: [coach]
      summary: Update a session
      parameters:
        - in: header
          name: x-coach-password
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                scheduledAt:
                  type: string
                durationMinutes:
                  type: integer
                service:
                  type: string
                notes:
                  type: string
                status:
                  type: string
      responses:
        "200":
          description: Session updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "404":
          description: Session not found
  /coach/invoices:
    get:
      operationId: getCoachInvoices
      tags: [coach]
      summary: List all invoices
      parameters:
        - in: header
          name: x-coach-password
          required: true
          schema:
            type: string
        - in: query
          name: clientId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createCoachInvoice
      tags: [coach]
      summary: Create a new invoice
      parameters:
        - in: header
          name: x-coach-password
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /coach/invoices/{id}:
    patch:
      operationId: updateCoachInvoice
      tags: [coach]
      summary: Update an invoice
      parameters:
        - in: header
          name: x-coach-password
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                amountCents:
                  type: integer
                status:
                  type: string
                dueDate:
                  type: string
                paymentMethod:
                  type: string
                notes:
                  type: string
      responses:
        "200":
          description: Invoice updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Invoice not found
    delete:
      operationId: deleteCoachInvoice
      tags: [coach]
      summary: Delete an invoice
      parameters:
        - in: header
          name: x-coach-password
          required: true
          schema:
            type: string
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice deleted
        "404":
          description: Invoice not found
  /plans/share/{token}:
    get:
      operationId: getSharedPlan
      tags: [plans]
      summary: Get a public coaching plan by share token
      parameters:
        - in: path
          name: token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public plan data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SharedPlan"
        "404":
          description: Plan not found
components:
  schemas:
...
    Booking:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        service:
          type: string
          enum: [private_lesson, group_session, stroke_clinic, video_analysis, package_5, package_10]
        preferredDate:
          type: string
        preferredTime:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [pending, confirmed, cancelled, approved, rejected]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - service
        - status
        - createdAt
...
    Client:
      type: object
      properties:
        id:
          type: integer
        bookingId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        neighborhood:
          type: string
        service:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [active, inactive]
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - status
        - createdAt
    CoachingPlan:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        title:
          type: string
        goal:
          type: string
        drills:
          type: string
        notes:
          type: string
        shareToken:
          type: string
        isPublic:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - shareToken
        - isPublic
        - createdAt
        - updatedAt
    CreatePlanRequest:
      type: object
      properties:
        title:
          type: string
        goal:
          type: string
        drills:
          type: string
        notes:
          type: string
        clientId:
          type: integer
        isPublic:
          type: boolean
      required:
        - title
    UpdatePlanRequest:
      type: object
      properties:
        title:
          type: string
        goal:
          type: string
        drills:
          type: string
        notes:
          type: string
        clientId:
          type: integer
        isPublic:
          type: boolean
    SharedPlan:
      type: object
      properties:
        title:
          type: string
        goal:
          type: string
        drills:
          type: string
        notes:
          type: string
        clientName:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - title
        - createdAt
        - updatedAt
    Session:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        bookingId:
          type: integer
        scheduledAt:
          type: string
        durationMinutes:
          type: integer
        service:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [scheduled, completed, cancelled]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - status
        - createdAt
        - updatedAt
    CreateSessionRequest:
      type: object
      properties:
        clientId:
          type: integer
        bookingId:
          type: integer
        scheduledAt:
          type: string
        durationMinutes:
          type: integer
        service:
          type: string
        notes:
          type: string
        status:
          type: string
    Invoice:
      type: object
      properties:
        id:
          type: integer
        clientId:
          type: integer
        sessionId:
          type: integer
        amountCents:
          type: integer
        status:
          type: string
          enum: [draft, sent, paid]
        dueDate:
          type: string
        paymentMethod:
          type: string
        notes:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - amountCents
        - status
        - createdAt
        - updatedAt
    CreateInvoiceRequest:
      type: object
      properties:
        clientId:
          type: integer
        sessionId:
          type: integer
        amountCents:
          type: integer
        status:
          type: string
        dueDate:
          type: string
        paymentMethod:
          type: string
        notes:
          type: string
      required:
        - amountCents