# lib/api-zod/src/generated/types/createInvoiceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Swim Coach API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateInvoiceRequest {
  clientId?: number;
  sessionId?: number;
  amountCents: number;
  status?: string;
  dueDate?: string;
  paymentMethod?: string;
  notes?: string;
}