# lib/api-zod/src/generated/types/createSessionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Swim Coach API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSessionRequest {
  clientId?: number;
  bookingId?: number;
  scheduledAt?: string;
  durationMinutes?: number;
  service?: string;
  notes?: string;
  status?: string;
}