# lib/api-zod/src/generated/types/invoice.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Swim Coach API specification
 * OpenAPI spec version: 0.1.0
 */
import type { InvoiceStatus } from "./invoiceStatus";

export interface Invoice {
  id: number;
  clientId?: number;
  sessionId?: number;
  amountCents: number;
  status: InvoiceStatus;
  dueDate?: string;
  paymentMethod?: string;
  notes?: string;
  createdAt: string;
  updatedAt: string;
}