# lib/api-zod/src/generated/types/session.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Swim Coach API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SessionStatus } from "./sessionStatus";

export interface Session {
  id: number;
  clientId?: number;
  bookingId?: number;
  scheduledAt?: string;
  durationMinutes?: number;
  service?: string;
  notes?: string;
  status: SessionStatus;
  createdAt: string;
  updatedAt: string;
}